feat: abstract the route
